Source code/ZyoFiSik-app/requirements.txt
absl-py==1.0.0
attrs==21.4.0
click==8.1.2
colorama==0.4.4
cycler==0.11.0
Flask==2.1.1
fonttools==4.31.2
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
gunicorn==20.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
Pillow==9.1.0
protobuf==3.20.0
pyparsing==3.0.7
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.1.1
zipp==3.7.0
